+++ words/random.py
@@ -1,9 +1,14 @@
 """Get a random line from a file."""
 
+from collections import namedtuple
+from functools import cached_property
 from pathlib import Path
 from random import choices
 
+from nltk import sent_tokenize
+
 from words import WordsError, bp  # noqa
+from words.object import Object
 
 
 class Random():
@@ -57,3 +62,85 @@ class RandomWord(Random):
         """Create the object."""
         path = self.ROOT / "wordlists" / "dictionary.txt"
         return super().__init__(file=path, **kwargs)
+
+
+class Sentence(namedtuple("Sentence", ["text", "source"])):
+    """Sentence class."""
+
+    @property
+    def title(self):
+        """Return the source title."""
+        return self.source.title
+
+
+class RandomParagraph():
+    """Randomly selected sentences from a paragraph of text."""
+
+    def __init__(self, text: str = None, source=None):
+        """Create the object."""
+        self.text = text and text.strip() or None
+        self.source = source
+
+    @property
+    def sentences(self) -> list[str]:
+        """Return a list of sentences."""
+        if not self.text:
+            return []
+
+        return sent_tokenize(self.text)
+
+    def get(self, count: int = 1) -> list[str]:
+        """Return a randomly selected list of sentences."""
+        return [Sentence(x, self) for x in choices(self.sentences, k=count)]
+
+    @property
+    def title(self):
+        """Return the source title."""
+        if not self.source:
+            return ""
+        return self.source.title
+
+
+class RandomProse():
+    """Randomly selected paragraphs from a prose file."""
+
+    def __init__(self, file: Path = None):
+        """Create the object."""
+        self.file = file
+
+    @property
+    def title(self) -> str:
+        """Return the title of the file."""
+        name = self.file.name
+        return name[:-4].replace("-", " ").title()
+
+    @cached_property
+    def text(self) -> str:
+        """Return the contents of the file."""
+        return self.file.read_text()
+
+    @cached_property
+    def paragraphs(self) -> list[str]:
+        """Return a list of paragraphs in the file."""
+        return self.text.split("\n\n")
+
+    def get(self, count: int = 1) -> list[str]:
+        """Get a randomly selected list of paragrpahs."""
+        return [RandomParagraph(x, self) for x in choices(self.paragraphs, k=count)]
+
+
+class RandomBooks(Object):
+    """Random selection of book files."""
+
+    ROOT = Path(__file__).parent.parent / "assets" / "prose"
+
+    _selection = []
+
+    @cached_property
+    def files(self) -> list[Path]:
+        """Return a list of text files that contain books."""
+        return list(self.ROOT.iterdir())
+
+    def get(self, count: int = 1) -> list[Path]:
+        """Return a list of randomly selected book files."""
+        return [RandomProse(x) for x in choices(self.files, k=count)]