+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New command: `words dict match` with options:
     - `--db`
     - `--strat`
+- New command: `words rand text` with options:
+    - `--unit`
+    - `--num`
 - Paginate all output when sensible.
 - Expand help text.
 - Add short help option `-h`.